+++ lua/plugins/custom/autoformat.lua
@@ -3,7 +3,6 @@ local setup = function()
 
 	conform.setup({
 		formatters_by_ft = {
-			lua = { "stylua" },
 			cpp = { "clang-format" },
 			c = { "clang-format" },
 		},